Changed body part system (SOs)
changed body part system from classes to scriptable objects. should help all members of the team be able to create their own body parts. :) Also changed the dictionary inside the player that used to hold the body parts to a parallel array. one is for BodyPartType enum and the other is the BodyPart SO class.